pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "algo-trading-engine"
version = "0.0.3"
description = "Options trading backtesting and prediction framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "yfinance>=1.0.0",
    "pandas>=2.2.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "hmmlearn>=0.3.0",
    "polygon-api-client>=1.14.5",
    "python-dotenv>=1.0.1",
    "tensorflow>=2.15.0",
    "tqdm>=4.66.0",
    "psutil>=5.9.0",
    "pytest>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
algo-backtest = "algo_trading_engine.backtest.main:main"
algo-paper-trade = "algo_trading_engine.prediction.recommend_cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["algo_trading_engine*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
pythonpath = ["src"]
